Integrate SDK with context-based engine
